Fix imsi info error when insert some china operator card.
Imsi only show mccmnc when the phone type is cdma phone since the
length of return from RIL is 0 but not NULL for mMin. if mMin is
NULL or length is 0, mPhone.getSubscriberId() will get NULL when
the phone type is cdma, hence the flow of get imsi will move on
to retrieve the imsi from IccRecords.

Change-Id: Idda892c736db046d8924bb45a5fbb2a5f2a6d248
CRs-Fixed: 2212375
Signed-off-by: spezi77 <spezi7713@gmx.net>